LS: Move all Salsa extensions to `lang::db` module
commit-id:d5d8c8b5
