+++ crates/cairo-lang-language-server/src/ide/completion/mod.rs
@@ -10,9 +10,8 @@ use tower_lsp::lsp_types::{CompletionParams, CompletionResponse, CompletionTrigg
 use tracing::debug;
 
 use self::completions::{colon_colon_completions, dot_completions, generic_completions};
+use crate::lang::db::{LsSemanticGroup, LsSyntaxGroup};
 use crate::lang::lsp::{LsProtoGroup, ToCairo};
-use crate::lang::semantic::LsSemanticGroup;
-use crate::lang::syntax::LsSyntaxGroup;
 
 mod completions;
