+++ crates/cairo-lang-language-server/src/lang/db/mod.rs
@@ -1,0 +1,5 @@
+pub use self::semantic::*;
+pub use self::syntax::*;
+
+mod semantic;
+mod syntax;
