+++ crates/cairo-lang-language-server/src/lang/mod.rs
@@ -1,4 +1,3 @@
+pub mod db;
 pub mod diagnostics;
 pub mod lsp;
-pub mod semantic;
-pub mod syntax;
